% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_postgres}
\alias{connect_postgres}
\title{Create a SQL DB conection}
\usage{
connect_postgres(alias = "rsqlauthandle")
}
\arguments{
\item{alias}{Str. Alias of db server. You need to set the connection before
create it}
}
\value{
PqConnection. Return the connection to db
}
\description{
Create a SQL DB conection for alias.
}
\examples{

\dontrun{
library(DBI)
con <- connect_postgres(alias='rsqlauthandle')
dbListTables(con)
}
}
