% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{set_credentials}
\alias{set_credentials}
\title{Set the credentials to connect to SQL DB.}
\usage{
set_credentials(
  alias = "rsqlauthandle",
  dialect = "postgresql",
  host = "localhost",
  port = "5432",
  user,
  db_name
)
}
\arguments{
\item{alias}{Service alias}

\item{dialect}{The dialect to config (...)}

\item{host}{Str. Host url of DB server.}

\item{port}{Str. Port number.}

\item{user}{Str. User.}

\item{db_name}{Str. Name of database to conect.}
}
\description{
Save dialect, host, port, db_name, user and app in config file. Save password
in system keyring. It will ask the db user password interactively.
}
