% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqlauthandle.R
\name{Sqlauth}
\alias{Sqlauth}
\title{Authorization class to manage authentication in SQL Databases}
\description{
Authorization class to manage authentication in SQL Databases

Authorization class to manage authentication in SQL Databases
}
\details{
It lets you handle and save SQL DB authentication configuration information.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alias}}{String identifier of the db conection.}

\item{\code{fileconf}}{The path to config file.}

\item{\code{config}}{A list with with the config options.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Sqlauth$new()}}
\item \href{#method-set_credentials}{\code{Sqlauth$set_credentials()}}
\item \href{#method-clone}{\code{Sqlauth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new sqlauth object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sqlauth$new(alias = "rsqlauthandle", resetfile = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alias}}{String identifier of the db conection.}

\item{\code{resetfile}}{Bool indicating if file will be reseted}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Sqlauth} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_credentials"></a>}}
\if{latex}{\out{\hypertarget{method-set_credentials}{}}}
\subsection{Method \code{set_credentials()}}{
Set the credentials to connect to SQL DB.

Save dialect, host, port, db_name, user and app in config file. Save password
in system keyring.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sqlauth$set_credentials(dialect, host, port, db_name, user, passwd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dialect}}{The dialect to config ...}

\item{\code{host}}{Str. Host url of DB server.}

\item{\code{port}}{Str. Port number.}

\item{\code{db_name}}{Str. Name of database to conect.}

\item{\code{user}}{Str. User.}

\item{\code{passwd}}{Str. Password}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sqlauth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
